+++ src/main/java/com/craftblocklock/client/config/CraftBlockLockModMenu.java
@@ -1,0 +1,11 @@
+package com.craftblocklock.client.config;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+
+public final class CraftBlockLockModMenu implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return CraftBlockLockConfigScreen::new;
+    }
+}